scripts/explosion_overlay/pyproject.toml
[project]
name = "explosion_overlay"
version = "1.0.0"
description = "Plays explosion as transparent fullscreen overlay using PNG sequence"
dependencies = [
    "Pillow>=9.0.0"
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"